# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APP SET V2",
        "description": "APP SET V2 Documentation",
        "contact": {
            "email": "alonso@setpay.mx"
        },
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/card/{pan_suffix}": {
            "get": {
                "tags": [
                    "Card Cloud"
                ],
                "summary": "Get card details by suffix",
                "description": "Retrieve card details using the last 8 digits of the card number",
                "operationId": "getCardDetails",
                "parameters": [
                    {
                        "name": "pan_suffix",
                        "in": "path",
                        "description": "Last 8 digits of the card number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "card_id": {
                                            "type": "string",
                                            "example": "01922a29-3edb-7283-895a-5aebfd93257d"
                                        },
                                        "card_external_id": {
                                            "type": "string",
                                            "example": "01922a29-4405-8d62-64d2-ceb079990479"
                                        },
                                        "card_type": {
                                            "type": "string",
                                            "example": "physical"
                                        },
                                        "brand": {
                                            "type": "string",
                                            "example": "MASTER"
                                        },
                                        "bin": {
                                            "type": "string",
                                            "example": "00000000"
                                        },
                                        "pan": {
                                            "type": "string",
                                            "example": "1122334455667788"
                                        },
                                        "client_id": {
                                            "type": "string",
                                            "example": "XX0000001"
                                        },
                                        "masked_pan": {
                                            "type": "string",
                                            "example": "XXXX XXXX XXXX 6694"
                                        },
                                        "balance": {
                                            "type": "string",
                                            "example": "0.00"
                                        },
                                        "clabe": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "BLOCKED"
                                        },
                                        "substatus": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "NORMAL"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "This card is already to use"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "setups": {
                                            "properties": {
                                                "Id": {
                                                    "type": "integer",
                                                    "example": 26302
                                                },
                                                "CardId": {
                                                    "type": "integer",
                                                    "example": 35115
                                                },
                                                "Status": {
                                                    "type": "string",
                                                    "example": "BLOCKED"
                                                },
                                                "StatusReason": {
                                                    "type": "string",
                                                    "example": "INITIAL_BLOCKED"
                                                },
                                                "Ecommerce": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "International": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "Stripe": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "Wallet": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "Withdrawal": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "Contactless": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "PinOffline": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "PinOnUs": {
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid card suffix format",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Invalid card suffix format. It must be 8 digits."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Unauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Error retrieving card details: {error_message}"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/card/{pan_suffix}/balance": {
            "get": {
                "tags": [
                    "Card Cloud"
                ],
                "summary": "Get card balance by suffix",
                "description": "Retrieve card balance using the last 8 digits of the card number",
                "operationId": "getCardBalance",
                "parameters": [
                    {
                        "name": "pan_suffix",
                        "in": "path",
                        "description": "Last 8 digits of the card number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card balance retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "balance": {
                                            "type": "string",
                                            "example": "100.00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid card suffix format",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Invalid card suffix format. It must be 8 digits."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Unauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Error retrieving card details: {error_message}"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}